Clean up more old stuff
